Refactor notification delivery
