+++ index.js
@@ -1,4 +1,4 @@
-require('dotenv').config();
+require("dotenv").config();
 
 const PORT = process.env.PORT || 8080;
