+++ lib/notification.js
@@ -1,7 +1,11 @@
 let db = require("./db");
+let { sendEmail, sendSMS } = require("./twilio");
 
 class Notification {
-    constructor(props) {
+    constructor(tandaID, props = {}) {
+        this.tandaID = tandaID;
+        this.delivered = false;
+
         Object.assign(this, props);
         this.ready = this.prepare();
     }
@@ -11,11 +15,31 @@ class Notification {
     async render() {
         return "Test notification please ignore";
     }
+
+    async deliver() {
+        if (!this.delivered) this.delivered = true;
+        else throw new Error("this notification has already been delivered");
+
+        let userIDs = (await db.getTandaByID()).members;
+        let users = await Promise.all(userIDs.map(db.getUserByID));
+
+        for (user of users) {
+            let { email, sms } = shouldNotifyWith(user, notification);
+
+            if (sms) {
+                await sendSMS(user.phone, notification.toString());
+            }
+
+            if (email) {
+                await sendEmail(user.email, notification.toString());
+            }
+        }
+    }
 }
 
 class ClaimCreatedNotification extends Notification {
     constructor(tandaID, claimantID) {
-        super({ tandaID, claimantID });
+        super(tandaID, { claimantID });
     }
 
     async prepare() {
@@ -30,7 +54,7 @@ class ClaimCreatedNotification extends Notification {
 
 class ClaimUpdatedNotification extends Notification {
     constructor(tandaID, claimantID) {
-        super({ tandaID, claimantID });
+        super(tandaID, { claimantID });
     }
 
     async prepare() {
@@ -45,7 +69,7 @@ class ClaimUpdatedNotification extends Notification {
 
 class ClaimApprovedNotification extends Notification {
     constructor(tandaID, claimantID, approverID) {
-        super({ tandaID, claimantID, approverID });
+        super(tandaID, { claimantID, approverID });
     }
 
     async prepare() {
@@ -62,7 +86,7 @@ class ClaimApprovedNotification extends Notification {
 
 class PremiumPaidNotification extends Notification {
     constructor(tandaID, payerID) {
-        super({ tandaID, payerID });
+        super(tandaID, { payerID });
     }
 
     async prepare() {
@@ -80,3 +104,12 @@ ClaimCreatedNotification.prototype.kind = "claim_created";
 ClaimUpdatedNotification.prototype.kind = "claim_updated";
 ClaimApprovedNotification.prototype.kind = "claim_approved";
 PremiumPaidNotification.prototype.kind = "premium_paid";
+
+function shouldNotifyWith(user, notification) {
+    let settings = user.notificationSettings;
+
+    let email = settings.email.includes(notification.kind);
+    let sms = settings.sms.includes(notification.kind);
+
+    return { email, sms };
+}
